scripts/requirements.txt
#aby zainstalować podane biblioteki:

mysqlclient
mysql-connector-python
pandas
sqlalchemy
pymysql
